backport r14091 from trunk: Fix flops computation in gemm

+++ examples/mult/xgemm.c
@@ -335,7 +335,7 @@ int main(int argc, char **argv)
 			task->handles[1] = starpu_data_get_sub_data(B_handle, 1, x);
 			task->handles[2] = starpu_data_get_sub_data(C_handle, 2, x, y);
 
-			task->flops = 2 * (xdim/nslicesx) * (ydim/nslicesy) * zdim;
+			task->flops = 2ULL * (xdim/nslicesx) * (ydim/nslicesy) * zdim;
 
 			ret = starpu_task_submit(task);
 			if (ret == -ENODEV)
@@ -357,8 +357,8 @@ int main(int argc, char **argv)
 
 	double timing = end - start;
 	double min;
-	double flops = 2.0*((unsigned long)niter)*((unsigned long)xdim)
-				*((unsigned long)ydim)*((unsigned long)zdim);
+	double flops = 2.0*((unsigned long long)niter)*((unsigned long long)xdim)
+		           *((unsigned long long)ydim)*((unsigned long long)zdim);
 
 	if (bound)
 	starpu_bound_compute(&min, NULL, 0);
